+++ CHANGELOG.rst
@@ -28,6 +28,7 @@ Deprecations:
 Changes:
 ^^^^^^^^
 
+- ``pem`` now ships with typing information that can be used by type checkers like `mypy <http://mypy-lang.org>`_.
 - PEM objects now have an ``obj.sha1_hexdigest`` property with the SHA-1 digest of the stored bytes  as a native string.
   This is the same digest as the one that is used by the PEM objects' ``__repr__``\ s.
 - PEM objects now have an ``obj.as_text()`` method that returns the PEM-encoded content as unicode, always.